+++ simbad/lattice/latticesearch.py
@@ -13,7 +13,6 @@ import datetime
 import logging
 import numpy as np
 import os
-import tempfile
 
 from simbad.lattice.latticescore import LatticeSearchResult
 import simbad.util.pdb_util
@@ -281,22 +280,16 @@ class LatticeSearch(object):
             msg = "Output directory does not exist: {0}".format(destination)
             raise ValueError(msg)
 
-        import gzip
-
         to_del = []
         for count, result in enumerate(self.results):
-            f_tmp_out = tempfile.NamedTemporaryFile(suffix='.pdb', delete=False)
             f_name = os.path.join(source, '{0}', 'pdb{1}.ent.gz').format(result.pdb_code[1:3].lower(),
                                                                          result.pdb_code.lower())
             f_name_out = os.path.join(destination, '{0}.pdb'.format(result.pdb_code))
             try:
-                with gzip.open(f_name, 'rb') as f_in:
-                    f_tmp_out.write(f_in.read())
-                struct = simbad.util.pdb_util.PdbStructure(f_tmp_out.name)
+                struct = simbad.util.pdb_util.PdbStructure()
+                struct.from_file(f_name)
                 struct.standardize()
                 struct.save(f_name_out)
-                os.unlink(f_tmp_out.name)
-
             except IOError:
                 logger.warning("Encountered problem copying PDB %s from %s - removing entry from list", result.pdb_code,
                                source)
@@ -332,16 +325,13 @@ class LatticeSearch(object):
 
         to_del = []
         for count, result in enumerate(self.results):
-            content = simbad.util.pdb_util.get_pdb_content(result.pdb_code)
-            if content:
-                f_tmp_out = tempfile.NamedTemporaryFile(suffix='.pdb', delete=False)
-                f_tmp_out.write(content)
-                struct = simbad.util.pdb_util.PdbStructure(f_tmp_out.name)
-                struct.standardize()
+            try:
                 f_name_out = os.path.join(destination, result.pdb_code + '.pdb')
+                struct = simbad.util.pdb_util.PdbStructure()
+                struct.from_pdb_code(result.pdb_code)
+                struct.standardize()
                 struct.save(f_name_out)
-                os.unlink(f_tmp_out.name)
-            else:
+            except RuntimeError:
                 logger.warning("Encountered problem downloading PDB %s - removing entry from list", result.pdb_code)
                 to_del.append(count)